tiny change LONG to INT

+++ TimeSheet_DB/src/main/java/com/swad/timesheet/timesheet_db/entity/Project.java
@@ -17,7 +17,7 @@ public class Project {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long pid;
+    private int pid;
     private String name;
